unauthorised customer shopping updated

+++ store/models.py
@@ -12,7 +12,7 @@ class Customer(models.Model):
         return self.name
 class Product(models.Model):
     name = models.CharField(max_length=200,null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False,null=True,blank=False)
     image = models.ImageField(null= True,blank= True)
     #image
@@ -57,7 +57,6 @@ class Order(models.Model):
         return total
       
     
- 
 class OrderItem(models.Model):
     product = models.ForeignKey(Product,on_delete=models.SET_NULL,null=True,blank=True)
     order = models.ForeignKey(Order,on_delete=models.SET_NULL,null=True,blank=True)
